+++ src/mocks/rules/validation/abstract/RuleAddressList/RuleAddressListInternal.sol
@@ -25,7 +25,7 @@ abstract contract RuleAddressListInternal {
      * If one of addresses already exist, there is no change for this address. The transaction remains valid (no revert).
      * @param listTargetAddresses an array with the addresses to list
      */
-    function _addAddressesToThelist(address[] calldata listTargetAddresses) internal {
+    function _addAddressesToThelist(address[] calldata listTargetAddresses) internal virtual {
         uint256 numAddressesListLocal = numAddressesList;
         for (uint256 i = 0; i < listTargetAddresses.length; ++i) {
             if (!list[listTargetAddresses[i]]) {
@@ -42,7 +42,7 @@ abstract contract RuleAddressListInternal {
      * The transaction remains valid (no revert).
      * @param listTargetAddresses an array with the addresses to remove
      */
-    function _removeAddressesFromThelist(address[] calldata listTargetAddresses) internal {
+    function _removeAddressesFromThelist(address[] calldata listTargetAddresses) internal virtual {
         uint256 numAddressesListLocal = numAddressesList;
         for (uint256 i = 0; i < listTargetAddresses.length; ++i) {
             if (list[listTargetAddresses[i]]) {
@@ -58,7 +58,7 @@ abstract contract RuleAddressListInternal {
      * If the address already exists, the transaction is reverted to save gas.
      * @param targetAddress The address to list
      */
-    function _addAddressToThelist(address targetAddress) internal {
+    function _addAddressToThelist(address targetAddress) internal virtual {
         if (list[targetAddress]) {
             revert Rulelist_AddressAlreadylisted();
         }
@@ -72,7 +72,7 @@ abstract contract RuleAddressListInternal {
      * @param targetAddress The address to remove
      *
      */
-    function _removeAddressFromThelist(address targetAddress) internal {
+    function _removeAddressFromThelist(address targetAddress) internal virtual {
         if (!list[targetAddress]) {
             revert Rulelist_AddressNotPresent();
         }
@@ -85,7 +85,7 @@ abstract contract RuleAddressListInternal {
      * @return Number of listed addresses
      *
      */
-    function _numberListedAddress() internal view returns (uint256) {
+    function _numberListedAddress() internal view virtual returns (uint256) {
         return numAddressesList;
     }
 
@@ -95,7 +95,7 @@ abstract contract RuleAddressListInternal {
      * @return True if the address is listed, false otherwise
      *
      */
-    function _addressIsListed(address _targetAddress) internal view returns (bool) {
+    function _addressIsListed(address _targetAddress) internal view virtual returns (bool) {
         return list[_targetAddress];
     }
 }